requirements.txt
strawberry-graphql
fastapi
uvicorn
asyncio
sqlalchemy
pydantic
apscheduler
asyncpg
redis
httpx
aiokafka
greenlet